sae_captioning_project/requirements.txt
# Core ML/DL
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
safetensors>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
datasets>=2.16.0
pyarrow>=14.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
kaleido>=0.2.1

# Analysis
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
umap-learn>=0.5.5

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
wandb>=0.16.0
einops>=0.7.0
jaxtyping>=0.2.25

# Arabic NLP
camel-tools>=1.5.0
pyarabic>=0.6.15

# Jupyter support
jupyter>=1.0.0
ipywidgets>=8.1.0
nbformat>=5.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
